Refactor filter bar into component and manager
