+++ src/main.ts
@@ -12,7 +12,7 @@ import { ScreenType } from './types/navigation';
 import { BottomsheetState, BottomsheetConfig } from './types/bottomsheet';
 import { SearchFlowManager } from './services/SearchFlowManager';
 import { BottomsheetManager } from './services/BottomsheetManager';
-import { MapSyncService, MapManager } from './services';
+import { MapSyncService, MapManager, FilterBarManager } from './services';
 import { DashboardScreen, DashboardScreenFactory } from './components/Screens/DashboardScreen';
 
 /**
@@ -28,6 +28,7 @@ class App {
   private dashboardScreen?: DashboardScreen;
   private searchFlowManager?: SearchFlowManager;
   private bottomsheetManager?: BottomsheetManager;
+  private filterBarManager?: FilterBarManager;
   private mapSyncService?: MapSyncService;
   private mapManager?: MapManager;
 
@@ -96,6 +97,9 @@ class App {
       window.innerHeight
     );
 
+    // Initialize FilterBarManager
+    this.filterBarManager = new FilterBarManager();
+
     // Initialize MapSyncService with dummy ref (will be updated by DashboardScreen)
     this.mapSyncService = new MapSyncService({ current: null });
 
@@ -109,7 +113,7 @@ class App {
    * Create dashboard screen using the modular component
    */
   private createDashboardScreen(): void {
-    if (!this.searchFlowManager || !this.bottomsheetManager) {
+    if (!this.searchFlowManager || !this.bottomsheetManager || !this.filterBarManager) {
       throw new Error('Services not initialized');
     }
 
@@ -118,6 +122,7 @@ class App {
       this.container,
       this.searchFlowManager,
       this.bottomsheetManager,
+      this.filterBarManager!,
       this.mapManager!
     );
